Updating pdb parser and removing styles parser
